+++ packages/babel-compiler/package.js
@@ -10,7 +10,7 @@ Package.describe({
 });
 
 Npm.depends({
-  'meteor-babel': '7.0.0-beta.38-2'
+  'meteor-babel': '7.0.0-beta.38-3'
 });
 
 Package.onUse(function (api) {