+++ packages/standard-minifier-css/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'standard-minifier-css',
-  version: '1.4.0',
+  version: '1.4.1',
   summary: 'Standard css minifier used with Meteor apps by default.',
   documentation: 'README.md'
 });